Update feature checkout

+++ includes/buyer/view_all_cart.h
@@ -1,4 +1,4 @@
 #ifndef VIEW_ALL_CART_H
 #define VIEW_ALL_CART_H
-int view_all_cart();
+float view_all_cart();
 #endif
